requirements.txt
fastapi
uvicorn
openai-whisper
googletrans==4.0.0-rc1